@@ -18,15 +18,17 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.15.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   provider: ^5.0.0
-  flutter_restart: ^0.0.3
+#  flutter_restart: ^0.0.3
   stop_watch_timer: ^1.3.1
-  flutter_animated_dialog: ^2.0.1
+  http: 0.13.4
+  localstorage: ^4.0.0+1
+#  flutter_animated_dialog: ^2.0.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
