@@ -1,12 +1,19 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
+import 'package:localstorage/localstorage.dart';
+import 'package:sudoku/consts.dart';
+import 'package:sudoku/models/UserModel.dart';
+import 'package:sudoku/screens/Community.dart';
 import 'package:sudoku/screens/CreatePuzzle.dart';
-import 'package:sudoku/screens/Game.dart';
+import 'package:sudoku/screens/CreatedGames.dart';
 import 'package:sudoku/screens/GameDifficulty.dart';
 import 'package:sudoku/screens/Statistics.dart';
 import 'package:sudoku/screens/themes.dart';
 
 class HomePageWidget extends StatefulWidget {
-  const HomePageWidget({Key key}) : super(key: key);
+  const HomePageWidget({Key? key}) : super(key: key);
 
   @override
   _HomePageWidgetState createState() => _HomePageWidgetState();
@@ -14,6 +21,31 @@ class HomePageWidget extends StatefulWidget {
 
 class _HomePageWidgetState extends State<HomePageWidget> {
   final scaffoldKey = GlobalKey<ScaffoldState>();
+  final LocalStorage storage = LocalStorage('localStorage');
+  late Future<UserModel> me;
+
+  Future<UserModel> fetchMe() async {
+    final token = storage.getItem('token');
+    final response = await http.get(Uri.parse(kUrl + '/user/me/'),
+        headers: {'Authorization': 'Bearer $token'});
+    if (response.statusCode == 200) {
+      var jsonResponse = json.decode(response.body);
+      final user = UserModel.fromJson(jsonResponse);
+      await storage.setItem('userId', user.id);
+      await storage.setItem('level', user.level);
+      await storage.setItem('xp', user.xp);
+      await storage.setItem('firstName', user.firstName);
+      return user;
+    } else {
+      throw Exception('Unexpected error occurred!');
+    }
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    me = fetchMe();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -39,25 +71,36 @@ class _HomePageWidgetState extends State<HomePageWidget> {
                       children: [
                         Padding(
                           padding: EdgeInsetsDirectional.fromSTEB(0, 20, 0, 0),
-                          child: Row(
-                            mainAxisSize: MainAxisSize.max,
-                            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                            children: [
-                              Padding(
-                                padding: EdgeInsetsDirectional.fromSTEB(
-                                    25, 0, 50, 0),
-                                child: Text(
-                                  'Level: 2',
-                                  style: kBodyText1White,
-                                ),
-                              ),
-                              SizedBox(
-                                child: Text(
-                                  'XP: 43',
-                                  style: kBodyText1White,
-                                ),
-                              ),
-                            ],
+                          child: FutureBuilder(
+                            future: me,
+                            builder: (context, snapshot) {
+                              if (snapshot.hasData) {
+                                UserModel? me = snapshot.data as UserModel?;
+                                return Row(
+                                  mainAxisSize: MainAxisSize.max,
+                                  mainAxisAlignment:
+                                      MainAxisAlignment.spaceEvenly,
+                                  children: [
+                                    Padding(
+                                      padding: EdgeInsetsDirectional.fromSTEB(
+                                          25, 0, 50, 0),
+                                      child: Text(
+                                        'Level: ${me?.level}',
+                                        style: kBodyText1White,
+                                      ),
+                                    ),
+                                    SizedBox(
+                                      child: Text(
+                                        'XP: ${me?.xp}',
+                                        style: kBodyText1White,
+                                      ),
+                                    ),
+                                  ],
+                                );
+                              } else {
+                                return Text('${snapshot.error}');
+                              }
+                            },
                           ),
                         ),
                       ],
@@ -129,7 +172,12 @@ class _HomePageWidgetState extends State<HomePageWidget> {
                           await Navigator.push(
                             context,
                             MaterialPageRoute(
-                              builder: (context) => StatsPageWidget(), //stats
+                              builder: (context) => StatsPageWidget(
+                                level: storage.getItem('level'),
+                                xp: storage.getItem('xp'),
+                                userId: storage.getItem('userId'),
+                                firstName: storage.getItem('firstName'),
+                              ),
                             ),
                           );
                         },
@@ -160,7 +208,8 @@ class _HomePageWidgetState extends State<HomePageWidget> {
                           await Navigator.push(
                             context,
                             MaterialPageRoute(
-                              builder: (context) => Container(), //profile page
+                              builder: (context) =>
+                                  CommunityWidget(), //profile page
                             ),
                           );
                         },
@@ -206,6 +255,11 @@ class _HomePageWidgetState extends State<HomePageWidget> {
                                 EdgeInsetsDirectional.fromSTEB(0, 30, 0, 0),
                             child: createPuzzleButton(),
                           ),
+                          Padding(
+                            padding:
+                                EdgeInsetsDirectional.fromSTEB(0, 30, 0, 0),
+                            child: createdGamesButton(),
+                          )
                         ],
                       ),
                     ),
@@ -252,6 +306,25 @@ class _HomePageWidgetState extends State<HomePageWidget> {
           );
         },
         child: Text('>     Create puzzle'),
+        style: kHomeButtonStyle,
+      ),
+    );
+  }
+
+  Widget createdGamesButton() {
+    return SizedBox(
+      width: 200,
+      height: 60,
+      child: ElevatedButton(
+        onPressed: () async {
+          await Navigator.push(
+            context,
+            MaterialPageRoute(
+              builder: (context) => CreatedGamesWidget(),
+            ),
+          );
+        },
+        child: Text('>     Created games'),
         style: kHomeButtonStyle,
       ),
     );
